+++ packages/core/react/src/index.ts
@@ -33,6 +33,8 @@ export * from "@jala-banyu/select";
 export * from "@jala-banyu/progress";
 export * from "@jala-banyu/alert";
 export * from "@jala-banyu/pagination";
+export * from "@jala-banyu/checkbox";
+export * from "@jala-banyu/table";
 /**
  * React Aria - Exports
  */