+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @jala-banyu/card
 
+## 4.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+  - @jala-banyu/shared-icons@1.2.1
+  - @jala-banyu/ripple@4.1.2
+
+## 4.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/shared-icons@1.2.0
+
+## 4.2.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+  - @jala-banyu/ripple@4.1.1
+
 ## 4.2.0
 
 ### Minor Changes