+++ packages/components/checkbox/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/checkbox
 
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+
 ## 1.1.0
 
 ### Minor Changes