+++ packages/components/chip/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @jala-banyu/chip
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+  - @jala-banyu/shared-icons@1.2.1
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+
 ## 1.1.0
 
 ### Minor Changes