+++ packages/components/code/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jala-banyu/code
 
+## 4.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+  - @jala-banyu/system-rsc@4.1.2
+
+## 4.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+  - @jala-banyu/system-rsc@4.1.1
+
 ## 4.1.0
 
 ### Minor Changes