+++ packages/components/image/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/image
 
+## 4.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+
+## 4.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+
 ## 4.1.0
 
 ### Minor Changes