+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/menu
 
+## 4.3.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+
+## 4.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+
 ## 4.3.0
 
 ### Minor Changes