+++ packages/components/modal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @jala-banyu/modal
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/shared-icons@1.2.1
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/shared-icons@1.2.0
+
 ## 1.0.0
 
 ### Major Changes