+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @jala-banyu/popover
 
+## 3.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.2
+  - @jala-banyu/button@4.1.2
+
+## 3.1.2
+
+### Patch Changes
+
+- fix popover
+
+## 3.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.4.1
+  - @jala-banyu/button@4.1.1
+
 ## 3.1.0
 
 ### Minor Changes