+++ packages/core/theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jala-banyu/theme
 
+## 1.4.2
+
+### Patch Changes
+
+- fix: table tooltip, turn off dark mode
+
+## 1.4.1
+
+### Patch Changes
+
+- updte component variant
+
 ## 1.4.0
 
 ### Minor Changes